pay-estimated-payouts: add a debt recovery mechanism
Backerei has an option --pay-estimated-rewards that allows to pay
rewards as soon as rights have been allocated, before the cycle
actually happens.

The downside is that actual rewards are almost always lower than
estimated rewards, since the estimates assume 100% uptime for
every baker and endorser.

We add an extra "pass" on the database which computes for every
delegator the difference between expected rewards and actual
rewards, on the cycle that has just finalized. This difference
is then withheld from the estimated rewards for the cycle for which
rights have just been calculated. The resulting amount is then paid
out to the delegate.

This behaviour kicks in after some payout cycles. Initially,
there is no withholding, since the delegation did not exist yet when
the current cycle's rights were calculated.

This results in the baker being able to accurately charge a fixed
fee for every cycle payout, based on the behaviour of other bakers
and endorsers, even though this charge is distributed across two
payouts.

This holds true as long as the address keeps delegating. If it stops
being a delegator, it can still walk away with some debt towards
the baker, but we do not believe this is significant.

All the relevant values are recorded in the databse, so a human-readable
explanation of the payout amounts can be generated.

Backerei assumes ideal baker operation. This change does not affect
that: own baker downtime is not taken into account. Own baker is assumed
always up.

We did not implement a specific configuration option for this
behaviour since we see little value in using the --pay-estiated-rewards
otherwise.

The code changes are mostly located in between checks for the
--pay-estimated-rewards option, so Cryptium's baker will likely not
be impacted by this change.